integrations/reddit-bot/requirements.txt
praw==7.7.1
python-dotenv==1.0.0
aiohttp==3.9.1
